PRSD-NONE: Renames pageobjects to pageObjects
