+++ src/test/kotlin/uk/gov/communities/prsdb/webapp/integration/pageObjects/components/Pagination.kt
@@ -1,8 +1,8 @@
-package uk.gov.communities.prsdb.webapp.integration.pageobjects.components
+package uk.gov.communities.prsdb.webapp.integration.pageObjects.components
 
 import com.microsoft.playwright.Locator
 import com.microsoft.playwright.Page
-import uk.gov.communities.prsdb.webapp.integration.pageobjects.pages.basePages.BasePage
+import uk.gov.communities.prsdb.webapp.integration.pageObjects.pages.basePages.BasePage
 import kotlin.reflect.KClass
 
 class Pagination<TPage : BasePage>(
