+++ src/test/kotlin/uk/gov/communities/prsdb/webapp/integration/pageObjects/pages/EditLaUserPage.kt
@@ -1,11 +1,11 @@
-package uk.gov.communities.prsdb.webapp.integration.pageobjects.pages
+package uk.gov.communities.prsdb.webapp.integration.pageObjects.pages
 
 import com.microsoft.playwright.Page
-import uk.gov.communities.prsdb.webapp.integration.pageobjects.components.BaseComponent.Companion.getButton
-import uk.gov.communities.prsdb.webapp.integration.pageobjects.components.BaseComponent.Companion.getHeading
-import uk.gov.communities.prsdb.webapp.integration.pageobjects.components.BaseComponent.Companion.getSubHeading
-import uk.gov.communities.prsdb.webapp.integration.pageobjects.components.Form
-import uk.gov.communities.prsdb.webapp.integration.pageobjects.pages.basePages.BasePage
+import uk.gov.communities.prsdb.webapp.integration.pageObjects.components.BaseComponent.Companion.getButton
+import uk.gov.communities.prsdb.webapp.integration.pageObjects.components.BaseComponent.Companion.getHeading
+import uk.gov.communities.prsdb.webapp.integration.pageObjects.components.BaseComponent.Companion.getSubHeading
+import uk.gov.communities.prsdb.webapp.integration.pageObjects.components.Form
+import uk.gov.communities.prsdb.webapp.integration.pageObjects.pages.basePages.BasePage
 
 class EditLaUserPage(
     page: Page,
