+++ src/test/kotlin/uk/gov/communities/prsdb/webapp/integration/pageObjects/pages/InviteNewLaUserSuccessPage.kt
@@ -1,0 +1,11 @@
+package uk.gov.communities.prsdb.webapp.integration.pageObjects.pages
+
+import com.microsoft.playwright.Page
+import uk.gov.communities.prsdb.webapp.integration.pageObjects.components.BaseComponent.Companion.getConfirmationPageBanner
+import uk.gov.communities.prsdb.webapp.integration.pageObjects.pages.basePages.BasePage
+
+class InviteNewLaUserSuccessPage(
+    page: Page,
+) : BasePage(page, "Invite sent") {
+    val confirmationBanner = getConfirmationPageBanner(page)
+}
