+++ src/test/kotlin/uk/gov/communities/prsdb/webapp/integration/pageObjects/pages/PageNotFoundPage.kt
@@ -1,7 +1,7 @@
-package uk.gov.communities.prsdb.webapp.integration.pageobjects.pages
+package uk.gov.communities.prsdb.webapp.integration.pageObjects.pages
 
 import com.microsoft.playwright.Page
-import uk.gov.communities.prsdb.webapp.integration.pageobjects.pages.basePages.BasePage
+import uk.gov.communities.prsdb.webapp.integration.pageObjects.pages.basePages.BasePage
 
 class PageNotFoundPage(
     page: Page,
