+++ src/test/kotlin/uk/gov/communities/prsdb/webapp/integration/pageObjects/pages/laUserRegistrationJourneyPages/NameFormPageLaUserRegistration.kt
@@ -1,9 +1,9 @@
-package uk.gov.communities.prsdb.webapp.integration.pageobjects.pages.laUserRegistrationJourneyPages
+package uk.gov.communities.prsdb.webapp.integration.pageObjects.pages.laUserRegistrationJourneyPages
 
 import com.microsoft.playwright.Page
 import uk.gov.communities.prsdb.webapp.constants.REGISTER_LA_USER_JOURNEY_URL
 import uk.gov.communities.prsdb.webapp.forms.steps.RegisterLaUserStepId
-import uk.gov.communities.prsdb.webapp.integration.pageobjects.pages.basePages.FormBasePage
+import uk.gov.communities.prsdb.webapp.integration.pageObjects.pages.basePages.FormBasePage
 
 class NameFormPageLaUserRegistration(
     page: Page,
