+++ src/test/kotlin/uk/gov/communities/prsdb/webapp/integration/pageObjects/pages/laUserRegistrationJourneyPages/SummaryPageLaUserRegistration.kt
@@ -1,9 +1,9 @@
-package uk.gov.communities.prsdb.webapp.integration.pageobjects.pages.laUserRegistrationJourneyPages
+package uk.gov.communities.prsdb.webapp.integration.pageObjects.pages.laUserRegistrationJourneyPages
 
 import com.microsoft.playwright.Page
 import com.microsoft.playwright.assertions.PlaywrightAssertions.assertThat
-import uk.gov.communities.prsdb.webapp.integration.pageobjects.components.BaseComponent.Companion.getHeading
-import uk.gov.communities.prsdb.webapp.integration.pageobjects.pages.basePages.BasePage
+import uk.gov.communities.prsdb.webapp.integration.pageObjects.components.BaseComponent.Companion.getHeading
+import uk.gov.communities.prsdb.webapp.integration.pageObjects.pages.basePages.BasePage
 
 class SummaryPageLaUserRegistration(
     page: Page,
