--- src/test/kotlin/uk/gov/communities/prsdb/webapp/integration/pageobjects/pages/InviteNewLaUserSuccessPage.kt
@@ -1,11 +1,0 @@
-package uk.gov.communities.prsdb.webapp.integration.pageobjects.pages
-
-import com.microsoft.playwright.Page
-import uk.gov.communities.prsdb.webapp.integration.pageobjects.components.BaseComponent.Companion.getConfirmationPageBanner
-import uk.gov.communities.prsdb.webapp.integration.pageobjects.pages.basePages.BasePage
-
-class InviteNewLaUserSuccessPage(
-    page: Page,
-) : BasePage(page, "Invite sent") {
-    val confirmationBanner = getConfirmationPageBanner(page)
-}
